refactor(signer): move fromSafeWebauthn to account/Safe/adapters feat(signer): validate publicKey coords are bigint in fromSafeWebauthn
